BindingResults are added. Validations are added. Ip config is added. ProductPhoto delete is edited.

+++ epay-cloud/src/main/java/com/lens/epay/service/ProductPhotoService.java
@@ -2,12 +2,12 @@ package com.lens.epay.service;
 
 import com.lens.epay.constant.ErrorConstants;
 import com.lens.epay.constant.HttpSuccessMessagesConstants;
-import com.lens.epay.exception.NotFoundException;
-import com.lens.epay.model.entity.Product;
+import com.lens.epay.exception.BadRequestException;
 import com.lens.epay.model.entity.ProductPhoto;
 import com.lens.epay.repository.ProductPhotoRepository;
 import com.lens.epay.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -28,19 +28,12 @@ public class ProductPhotoService {
     @Autowired
     private ProductPhotoRepository repository;
 
+    @Autowired
+    private ProductService productService;
+
     @Transactional
     public String uploadProductPhoto(MultipartFile file, UUID productId) {
-        Product product = productRepository.findById(productId).orElse(null);
-        if (product == null) {
-            throw new NotFoundException(ErrorConstants.PRODUCT_NOT_EXIST);
-        }
-        ProductPhoto photo;
-        if (repository.existsByProductId(productId)) {
-            photo = repository.findProductPhotoByProductId(productId);
-        } else {
-            photo = new ProductPhoto();
-            photo.setProduct(product);
-        }
+        ProductPhoto photo = repository.findProductPhotoByProductId(productId).orElse(new ProductPhoto(productService.fromIdToEntity(productId)));
         try {
             photo.setFile(file.getBytes());
         } catch (IOException e) {
@@ -52,16 +45,19 @@ public class ProductPhotoService {
 
     @Transactional
     public byte[] getPhoto(UUID productId) {
-        ProductPhoto photo = repository.findProductPhotoByProductId(productId);
+        ProductPhoto photo = repository.findProductPhotoByProductId(productId).orElse(null);
         if (photo == null) {
-            return null;
+            throw new BadRequestException(ErrorConstants.ID_IS_NOT_EXIST);
         }
         return photo.getFile();
     }
 
     @Transactional
+    @Modifying
     public void deletePhotoByProductId(UUID productId) {
-        repository.deleteProductPhotoByProductId(productId);
+        if (repository.existsByProductId(productId)) {
+            repository.deleteProductPhotoByProductId(productId);
+        }
     }
 
     @Transactional
